fix: keep node-only devnet config out of browser bundles
get-rpc-url.ts statically imported node:fs/os/path for devnet support,
which Rollup/Vite cannot bundle for the browser. The browser only reaches
it via a guarded lazy import in resolveChainFromRpc, but bundlers still
pull in the lazy target and choke on the node built-ins.

Move the devnet config loading into its own module and provide a browser
stub, swapped in via the package.json "browser" field. resolveChainFromRpc's
existing try/catch handles the stub's throw on the (browser-only) devnet
probe path.

This restores browser/Rollup support so consumers no longer need their own
stub or bundler plugin to depend on filecoin-pin.
